Renamed GBFSaveGame to GBFSettingsShared
