+++ Source/GameBaseFramework/Private/GameBaseFrameworkPCH.h
@@ -2,4 +2,4 @@
 
 #include "Engine/GBFGameInstance.h"
 #include "GameBaseFrameworkDeveloperSettings.h"
-#include "GameFramework/GBFSaveGame.h"
+#include "GameFramework/GBFSettingsShared.h"
